Fixing capture error
